requirements.txt
# Core dependencies
flask
requests
python-dotenv
supabase-py
